update get apphost info
